requirements.txt
fastapi
supabase
pydantic
cerebras-cloud-sdk